feat: Se terminan de proteger todos los metodos por businessuuid.
Se mueve parte de la logica directamente al jwtservice.

+++ src/main/java/com/example/demo/services/IngredientService.java
@@ -3,13 +3,11 @@ package com.example.demo.services;
 import com.example.demo.dto.IngredientRequest;
 import com.example.demo.dto.mapper.IngredientMapper;
 import com.example.demo.entity.Ingredient;
-import com.example.demo.entity.Product;
-import com.example.demo.entity.User;
 import com.example.demo.jwt.JwtService;
 import com.example.demo.repository.IngredientRepository;
+import jakarta.annotation.PostConstruct;
 import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityNotFoundException;
-import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
@@ -25,21 +23,28 @@ public class IngredientService {
   private final ProductService productService;
   private static final String NOT_FOUND_TEXT = "Ingredient not found";
   private final JwtService jwtService;
-  private final HttpServletRequest request;
   private final UserService userService;
-  private final BussinesService bussinesService;
+  private String bussinesUUid;
 
+  @PostConstruct
+  public void init() {
+    try {
+      this.bussinesUUid = userService.findUserbyUser(jwtService.getUsername())
+          .getBusinessUuid();
+    } catch (Exception exception) {
+      this.bussinesUUid = "";
+    }
+  }
 
   public Ingredient getIngredientById(Long id) {
-    User user = userService.findUserbyUser(jwtService.getUsername(request));
-    Optional<Ingredient> ingredient = ingredientRepository.findByIdAndBusinnes(id, user.getBusinessUuid());
+    Optional<Ingredient> ingredient = ingredientRepository.findByIdAndBusinnesUUid(id,
+        bussinesUUid);
     return ingredient.orElseThrow(() -> new EntityNotFoundException(NOT_FOUND_TEXT));
   }
 
   public Ingredient getIngredientByProductId(Long productId) {
-    Product product = new Product();
-    product.setId(productId);
-    Optional<Ingredient> ingredient = ingredientRepository.findByProductId(product);
+    Optional<Ingredient> ingredient = ingredientRepository.findByProductIdAndBusinnesUuid(productId,
+        bussinesUUid);
     return ingredient.orElseThrow(() -> new EntityNotFoundException(NOT_FOUND_TEXT));
   }
 
@@ -48,29 +53,33 @@ public class IngredientService {
     Ingredient ingredient = ingredientMapper.toEntity(
         ingredientRequest);
     try {
-//      productService.getProductById(Math.toIntExact(ingredient.getProduct().getId()));
+      if (productService.getProductById(ingredient.getProduct().getId()) != null) {
+        return ingredientRepository.save(ingredient);
+      }
     } catch (NoSuchElementException exception) {
       throw new EntityNotFoundException("Product not found");
     }
-    return ingredientRepository.save(ingredient);
+    return null;
   }
 
   public Ingredient updateIngredient(Ingredient ingredient) {
-    ingredientRepository.findById(ingredient.getId())
+    ingredientRepository.findByIdAndBusinnesUUid(ingredient.getId(), bussinesUUid)
         .orElseThrow(() -> new EntityNotFoundException(NOT_FOUND_TEXT));
     return ingredientRepository.save(ingredient);
   }
 
   public void deleteIngredient(Long id) {
-    ingredientRepository.deleteById(id);
+
+    if (ingredientRepository.findByIdAndBusinnesUUid(id, bussinesUUid).isPresent()) {
+      ingredientRepository.deleteById(id);
+    }
     if (ingredientRepository.existsById(id)) {
       throw new EntityExistsException("Ingredient not deleted.");
-
     }
   }
 
   public List<Ingredient> getOrders() {
-    List<Ingredient> ingredients = ingredientRepository.findAll();
+    List<Ingredient> ingredients = ingredientRepository.findAllByBusinessUuid(bussinesUUid);
     if (ingredients.isEmpty()) {
       throw new EntityNotFoundException("Orders not found");
     }
